requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
ansicolors==1.1.8
anthropic==0.52.0
anyio==4.10.0
asn1crypto==1.5.1
astrapy==1.4.2
astunparse==1.6.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
blis==1.3.0
boto3==1.40.11
botocore==1.40.11
branca==0.8.1
bson==0.5.10
build==1.3.0
cachetools==5.5.2
cassandra-driver==3.29.2
cassio==0.1.10
catalogue==2.0.10
causal-learn==0.1.4.3
certifi==2025.8.3
cffi==1.15.0
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.3.0
cloudpathlib==0.21.1
cloudpickle==3.0.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
cryptography==45.0.6
curl_cffi==0.11.1
cycler==0.12.1
cymem==2.0.11
daily-python==0.19.3
dataclasses-json==0.5.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
entrypoints==0.4
et_xmlfile==2.0.0
eventlet==0.33.3
Faker==37.5.3
fastapi==0.110.1
fastjsonschema==2.21.1
favicon==0.7.0
feedparser==6.0.11
filelock==3.18.0
Flask==2.3.3
flask-cors==5.0.1
Flask-SocketIO==5.3.6
flatbuffers==25.2.10
folium==0.20.0
fonttools==4.55.8
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
futures==3.0.5
fuzzywuzzy==0.18.0
gast==0.4.0
geographiclib==2.0
geomet==0.2.1.post1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hf-xet==1.1.5
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.13.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.0.92
language_data==1.3.0
Levenshtein==0.27.1
libclang==18.1.1
logging==0.4.9.6
lxml==6.0.0
marisa-trie==1.3.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.7.2
mdurl==0.1.2
mediapipe==0.10.7
ml_dtypes==0.5.3
mmh3==5.2.0
momentchi2==0.1.8
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.11
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.42.1
nbclient==0.10.2
nbformat==5.10.4
networkx==3.5
nltk==3.9.1
numexpr==2.11.0
numpy==1.24.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==0.28.1
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.0.3
papermill==2.6.0
patsy==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.14.10
pillow==11.1.0
pip-chill==1.0.3
platformdirs==4.3.8
plotly==6.3.0
posthog==5.4.0
preshed==3.0.10
prometheus_client==0.22.1
propcache==0.3.2
protobuf==3.20.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.2
pycparser==2.21
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
pymongo==4.13.2
PyMySQL==1.0.2
pyOpenSSL==25.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.3
python-Levenshtein==0.27.1
python-socketio==5.9.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.12.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
selenium==4.32.0
sentence-transformers==5.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.17.1
snowflake-snowpark-python==1.36.0
sortedcontainers==2.4.0
sounddevice==0.5.3
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
srsly==2.5.1
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.37.2
statsmodels==0.14.4
streamlit==1.45.1
streamlit-agraph==0.0.45
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.6
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-option-menu==0.4.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==3.1.0
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tigramite==5.2.8.2
tiktoken==0.10.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@6dfd561d9cd722dfc09f702355518c6d09b9b4e3
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.5.0
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.5.0
uuid6==2024.1.12
uvicorn==0.29.0
uvloop==0.21.0
vaderSentiment==3.3.2
validators==0.35.0
vapi-python==0.1.9
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.3
wsproto==1.2.0
xgboost==3.0.0
xyzservices==2025.4.0
yarl==1.20.1
yfinance==0.2.61
zipp==3.23.0
zstandard==0.23.0
